requirements.txt
alembic==1.7.5
amqp==5.0.9
anyio==3.4.0
asgiref==3.4.1
attrs==21.4.0
autoflake==1.4
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.2
dnspython==2.1.0
docopt==0.6.2
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
iniconfig==1.1.1
isort==5.10.1
jose==1.0.0
kombu==5.2.3
Mako==1.1.6
MarkupSafe==2.0.1
mypy==0.930
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.24
py==1.11.0
pyasn1==0.4.8
pydantic==1.9.0
pyflakes==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
requests==2.27.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
sqlalchemy-stubs==0.4
starlette==0.16.0
toml==0.10.2
tomli==1.2.3
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
vine==5.0.0
wcwidth==0.2.5
yarg==0.1.9
zipp==3.7.0
